requirements.txt
pycryptodome
gmpy2
primefac==2.0.12